feat(admin): add edit page for entity type categories
Introduce EditEntityTypeCategoryPage component to enable editing and
deleting entity type categories within directories. This page includes
form inputs for name and label, a save button to update the category,
and a confirmation modal for safe deletion. Remove redundant client-side
code from the page.tsx to streamline server-side data fetching and
rendering. These changes improve category management UX and maintain
cleaner code separation.
